+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Machavaram 2013 interleukin-6 / CYP3A4 suppression ([doi:10.1038/clpt.2013.79](https://doi.org/10.1038/clpt.2013.79)) -- adults with rheumatoid arthritis, bone marrow transplant or surgical trauma.
+- Add Mandema 2011 anticoagulants dose-response meta-analysis ([doi:10.1038/clpt.2011.232](https://doi.org/10.1038/clpt.2011.232)) -- adults undergoing total hip or knee replacement surgery.
 
 - Add Siebinga 2023 [68Ga]Ga- and [177Lu]Lu-HA-DOTATATE ([doi:10.1186/s40658-023-00565-4](https://doi.org/10.1186/s40658-023-00565-4)) - adults with neuroendocrine tumors.